HW2/requirements.txt
aiogram==3.*
python_dotenv==1.0.1
requests==2.32.3